fix: race condition in insert_auto and duplicate handling in par_insert_batch
- insert_auto now assigns ID and inserts atomically under a single write
  lock, preventing concurrent calls from receiving the same ID
- par_insert_batch now removes stale hash entries when a duplicate ID is
  encountered, matching the behavior of the regular insert method
- metrics avg_query_time_us now uses float division to avoid precision loss

+++ src/metrics.rs
@@ -49,7 +49,7 @@ impl MetricsCollector {
             query_count,
             insert_count: self.insert_count.load(Ordering::Relaxed),
             avg_query_time_us: if query_count > 0 {
-                (total_query_time_ns / query_count) as f64 / 1000.0
+                total_query_time_ns as f64 / query_count as f64 / 1000.0
             } else {
                 0.0
             },
